services/api-gateway/requirements.txt
# API Gateway Dependencies

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP Client for proxying
httpx==0.25.1

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Environment
python-dotenv==1.0.0
